% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{bic}
\alias{bic}
\title{Calculate the Bayesian Information Criteron for 0 to b breaks}
\usage{
bic(formula, data, b, ssr)
}
\arguments{
\item{formula}{formula y~x object}

\item{data}{data}

\item{b}{number of breaks}

\item{ssr}{ssr table}
}
\description{
Calculate the Bayesian Information Criteron for 0 to b breaks
}
