% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{conf_int}
\alias{conf_int}
\title{Calculate Confidence intervals for 2.5\% and 5\% significance}
\usage{
conf_int(formula, data, from, to, autocorrelation = FALSE)
}
\arguments{
\item{formula}{formula y~x object}

\item{data}{data}

\item{from}{segment position from}

\item{to}{segment position to}

\item{autocorrellation}{autocorrelation present?}
}
\description{
Calculate Confidence intervals for 2.5\% and 5\% significance
}
