% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{dmax_stats}
\alias{dmax_stats}
\title{Double maximum test for up to the number of breaks}
\usage{
dmax_stats(formula, data, breaks, ssr, conf = 0.95, trim = 0.1)
}
\arguments{
\item{formula}{y~x formula object}

\item{data}{cci dataframe}

\item{breaks}{number of 1:breaks to test for}

\item{ssr}{SSR Dataframe}

\item{conf}{confidence interval}

\item{trim}{trimming 0.05 - 0.25}
}
\description{
Double maximum test for up to the number of breaks
}
