% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{kpi}
\alias{kpi}
\title{Return KPI for Construction Cost Index data}
\usage{
kpi(cci = cci())
}
\arguments{
\item{cci}{Data as defined in the cci() function}
}
\description{
Return KPI for Construction Cost Index data
}
