% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{lstats}
\alias{lstats}
\title{F-test for l+1|l breaks}
\usage{
lstats(formula, data, l, ssr, conf = 0.95, trim = 0.1, autocorrelation = FALSE)
}
\arguments{
\item{formula}{y~x formula object}

\item{data}{cci dataframe}

\item{ssr}{SSR Dataframe}

\item{conf}{confidence interval}

\item{trim}{trimming 0.05 - 0.25}

\item{autocorrelation}{does the data contain autocorrelation?}

\item{breaks}{number of 1:breaks to test for}
}
\description{
F-test for l+1|l breaks
}
