% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{lwz}
\alias{lwz}
\title{Calculate the modified Schwarz criterion for 0 to b breaks}
\usage{
lwz(formula, data, b, ssr)
}
\arguments{
\item{formula}{formula y~x object}

\item{data}{data}

\item{b}{number of breaks}

\item{ssr}{ssr table}
}
\description{
Calculate the modified Schwarz criterion for 0 to b breaks
}
