% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{min_ssr}
\alias{min_ssr}
\title{Dynamic programming algorithm for the global minimum of SSR for a number of breaks}
\usage{
min_ssr(breaks, ssr, init = 0)
}
\arguments{
\item{breaks}{number of breaks}

\item{ssr}{ssr table}

\item{init}{begin of first observation - 1}
}
\description{
Dynamic programming algorithm for the global minimum of SSR for a number of breaks
}
