% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_annotate}
\alias{plot_annotate}
\title{annotate the plot}
\usage{
plot_annotate(label, x, y, xpos, ypos, hjust)
}
\arguments{
\item{label}{label}

\item{x, y}{pos of data}

\item{xpos, ypos}{pos of label}
}
\description{
annotate the plot
}
