% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_breaks}
\alias{plot_breaks}
\title{Plot breaks in data}
\usage{
plot_breaks(data, segments)
}
\arguments{
\item{data}{cci dataframe}

\item{segments}{segments dataframe with start, end and id}
}
\description{
Plot breaks in data
}
