% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_confidence}
\alias{plot_confidence}
\title{draw confidence interval}
\usage{
plot_confidence(x, y)
}
\arguments{
\item{x}{lower and upper bound of x}

\item{y}{y position}
}
\description{
draw confidence interval
}
