% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{rec_res}
\alias{rec_res}
\title{calculate the squared recurive residual in a segment}
\usage{
rec_res(formula, start, end, data)
}
\arguments{
\item{formula}{y~x formula object}

\item{start}{start of segment}

\item{end}{end of segment}

\item{data}{dataframe of observations}
}
\description{
calculate the squared recurive residual in a segment
}
