% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{ssr}
\alias{ssr}
\title{calculate sum of squared residuals}
\usage{
ssr(formula, data, trim = 0.1)
}
\arguments{
\item{data}{dataframe of observations}

\item{trim}{trimming for minimal length, either percentage or absolute value}

\item{forumla}{x~y formulate object}
}
\description{
calculate sum of squared residuals
}
\details{
Calculates all SSR under consideration of the minimal length requirement
}
