% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zbar}
\alias{zbar}
\title{Calculate \overline(Z) matrix}
\usage{
zbar(z, bpos)
}
\arguments{
\item{z}{regressors as matrix}

\item{bpos}{vector of positions where break occurs after}
}
\description{
Calculate \overline(Z) matrix
}
