% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{dmax_stats}
\alias{dmax_stats}
\title{Double maximum test for up to the number of breaks}
\usage{
dmax_stats(formula, data, breaks, ssr, sig = "5\%")
}
\arguments{
\item{formula}{y~x formula object}

\item{data}{cci dataframe}

\item{breaks}{number of 1:breaks to test for}

\item{ssr}{SSR Dataframe}

\item{sig}{significance level}
}
\description{
Double maximum test for up to the number of breaks
}
