% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{fstats}
\alias{fstats}
\title{Calculate F-statistic value}
\usage{
fstats(formula, data, breaks, ssr)
}
\arguments{
\item{formula}{y~x formula object}

\item{data}{cci dataframe}

\item{breaks}{number of breaks to test for}

\item{ssr}{SSR Dataframe}
}
\description{
Calculate F-statistic value
}
