% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{rec_res}
\alias{rec_res}
\title{calculate the squared recursive residuals in a segment}
\usage{
rec_res(formula, from, to, data, h)
}
\arguments{
\item{formula}{y~x formula object}

\item{from}{from of segment}

\item{to}{to of segment}

\item{data}{dataframe of observations}
}
\description{
calculate the squared recursive residuals in a segment
}
