% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{seq_test}
\alias{seq_test}
\title{Sequential testing for another break up to m}
\usage{
seq_test(formula, data, m, ssr, conf = 0.95, trim = 0.1)
}
\arguments{
\item{formula}{y~x formula object}

\item{data}{cci dataframe}

\item{ssr}{SSR Dataframe}

\item{conf}{confidence interval}

\item{trim}{trimming 0.05 - 0.25}

\item{breaks}{number of 1:breaks to test for}
}
\description{
Sequential testing for another break up to m
}
