% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.R
\name{valid_segm}
\alias{valid_segm}
\title{Function to decide on valid segments in series}
\usage{
valid_segm(from, to, length, n_breaks, min_length)
}
\arguments{
\item{from}{begin of segment}

\item{to}{to of segment}

\item{length}{length of total observation period}

\item{n_breaks}{number of breaks in data}

\item{min_length}{minimum length of segment}
}
\description{
Function to decide on valid segments in series
}
